+++ components/common/index.ts
@@ -2,7 +2,8 @@ import NavBar from "./NavBar/NavBar";
 import Pagination from "./Pagination/Pagination";
 import SearchBar from "./SearchBar/SearchBar";
 import StatusChip from "./StatusChip/StatusChip";
+import CustomInput from "./CustomInput/CustomInput";
 
 export {
-  NavBar, Pagination, SearchBar, StatusChip,
+  NavBar, Pagination, SearchBar, StatusChip, CustomInput,
 };
