+++ public/sitemap-0.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:mobile="http://www.google.com/schemas/sitemap-mobile/1.0" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1">
+<url><loc>https://www.thejulge.com/api/sample</loc><lastmod>2023-07-09T02:48:00.155Z</lastmod><changefreq>daily</changefreq><priority>1</priority></url>
+<url><loc>https://www.thejulge.com</loc><lastmod>2023-07-09T02:48:00.155Z</lastmod><changefreq>daily</changefreq><priority>1</priority></url>
+<url><loc>https://www.thejulge.com/kenny</loc><lastmod>2023-07-09T02:48:00.155Z</lastmod><changefreq>daily</changefreq><priority>1</priority></url>
+</urlset>
